requirements.txt
torch>=2.0
transformers>=4.35
sentence-transformers>=2.2
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4
numpy>=1.23
tqdm>=4.66


